pyproject.toml
[project]
name = "image-activity"
version = "0.1.0"
description = "Analyze screenshot, saved-image, and camera-photo activity patterns."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "cairosvg>=2.8",
  "matplotlib>=3.8",
  "pandas>=2.2",
  "pillow>=10.0",
  "pytesseract>=0.3.10",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "seaborn>=0.13",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=3.8",
  "ruff>=0.8.0",
]

[project.scripts]
activity = "activity:main"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["activity.py", "plots.py"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
